+++ framework/domain/src/main/java/is/codion/framework/domain/entity/DefaultEntityDefinition.java
@@ -312,8 +312,8 @@ final class DefaultEntityDefinition implements EntityDefinition, Serializable {
   }
 
   @Override
-  public OrderBy orderBy() {
-    return orderBy;
+  public Optional<OrderBy> orderBy() {
+    return Optional.ofNullable(orderBy);
   }
 
   @Override
@@ -322,8 +322,8 @@ final class DefaultEntityDefinition implements EntityDefinition, Serializable {
   }
 
   @Override
-  public SelectQuery selectQuery() {
-    return selectQuery;
+  public Optional<SelectQuery> selectQuery() {
+    return Optional.ofNullable(selectQuery);
   }
 
   @Override