+++ framework/model/src/main/java/is/codion/framework/model/EntityTableConditionModel.java
@@ -48,7 +48,7 @@ public interface EntityTableConditionModel<C extends Attribute<?>> extends Table
    * Sets the search condition values of the condition model associated with {@code attribute}.
    * Enables the condition model in case {@code values} is non-empty or disables it if {@code values is empty}.
    * @param attribute the attribute
-   * @param values the search condition values
+   * @param values the search condition values, an empty Collection for none
    * @param <T> the value type
    * @return true if the search state changed as a result of this method call, false otherwise
    */