+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/text/TextFieldPanel.java
@@ -70,6 +70,9 @@ public final class TextFieldPanel extends JPanel {
    * The available keyboard shortcuts.
    */
   public enum KeyboardShortcut {
+    /**
+     * Displays a text area for longer text input.
+     */
     DISPLAY_TEXT_AREA
   }